Vendor git pre-receive hook
Instead of fetching it from the Internet. This will protect us against
GitHub being down or the script changing and being inconsistent across
environments.

It also means that we're no longer using inline shell script. Which Ansible
was giving us a warning about and we wouldn't have been capturing exit codes
if anything but the last command failed.

The file is taken from this upstream:

https://github.com/tsuru/tsuru/blob/e0699f287dce9a48e3b727a939bb179df5e7e98c/misc/git-hooks/pre-receive.archive-server

This change is a noop when applied to an existing machine. I'm still able to
deploy an app when this is applied to a new Gandalf machine.