Setup queue:mongo-url and queue:mongo-database
The snapshot version of tsuru requires this configuration to be set,
despite the documentation still describes it as optional. In
the latest versions of tsuru (>0.11.3) this queue is
used for different internal tasks, included provisioner.

We need to populate these variables in order to install the latest
snapshot version of tsuru which includes support for vulcand.